requirements.txt
numpy
pandas
matplotlib
seaborn

tensorflow
keras
scikit-learn
nltk
textblob
statsmodels
tslearn

pyyaml
loguru

joblib

plotly

Flask
Flask-Cors
jinja2

notebook
jupyter
pykernel

dvc
scipy
pytest

black
isort
mypy
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
types-pyYAML
-e .